separate legacy and new subscription type repos
