+++ notifications/src/main/java/dk/shape/games/notifications/repositories/NotificationsServiceBase.kt
@@ -9,33 +9,26 @@ import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.http.*
 
-interface NotificationsService : NotificationsServiceBase {
+interface NotificationsServiceBase {
 
     @Headers("Content-Type: application/json")
-    @GET("/api/v1/subscriptions/{deviceId}")
-    suspend fun getSubscriptions(
-        @Path("deviceId") deviceId: String
-    ): List<Subscription>
+    @PUT("/api/v1/subscriptions")
+    suspend fun updateSubscriptions(
+        @Body request: SubscribeRequest
+    )
 
     @Headers("Content-Type: application/json")
-    @GET("/api/v1/subscriptions/{deviceId}?subject_type={subject_type}")
-    suspend fun getSubscriptionsForType(
-        @Path("deviceId") deviceId: String,
-        @Query("subject_type") subjectType: SubjectType
-    ): List<Subscription>
-
-    @Headers("Content-Type: application/json")
-    @GET("/api/v1/subscriptions/{deviceId}?showAll=true")
-    suspend fun getAllSubscriptions(
-        @Path("deviceId") deviceId: String
-    ): List<Subscription>
+    @POST("/api/v1/register")
+    suspend fun register(
+        @Body request: RegistrationRequest
+    )
 
     companion object {
 
         private val contentType = "application/json".toMediaType()
 
         @JvmStatic
-        fun create(baseUrl: String, httpClient: OkHttpClient): NotificationsService {
+        fun create(baseUrl: String, httpClient: OkHttpClient): NotificationsServiceBase {
             val retrofit = Retrofit.Builder()
                 .addConverterFactory(
                     Json {
@@ -47,7 +40,7 @@ interface NotificationsService : NotificationsServiceBase {
                 .client(httpClient)
                 .build()
 
-            return retrofit.create(NotificationsService::class.java)
+            return retrofit.create(NotificationsServiceBase::class.java)
         }
     }
 }
